gateway-api: extract listener merge and conflict handling
Move Gateway and ListenerSet listener status handling into
one API entrypoint `ListenerStatusManager.SetListenerStatuses`.

This allows moving listener conflict detection, parent-ref matching, and
listener filtering out of `gateway_reconcile.go`, adds a combined
`SetListenerStatuses` flow, and returns the valid merged listeners used
for ingestion.

It also keeps the listener-status helpers/tests co-located and removes
the old reconcile-only helper code.
